CORE: SERVICES: fix compilation error with -Werror=absolute-value
Transfer as signed integer in __abf_os_abs function.
Build error log:
  taking the absolute value of unsigned type ‘u_int32_t’
  {aka ‘unsigned int’} has no effect [-Werror=absolute-value]

+++ CORE/SERVICES/DFS/src/dfs_process_radarevent.c
@@ -388,7 +388,7 @@ static void dfs_process_dc_pulse(struct ath_dfs *dfs, struct dfs_event *event,
                                (rf->rf_dl.dl_elems[i].de_time), min_pri);
                         deviation = __adf_os_abs(min_pri *
                                         miss_pulse_number -
-                                        rf->rf_dl.dl_elems[i].de_time);
+                                        (int)rf->rf_dl.dl_elems[i].de_time);
                         if (deviation > miss_pulse_number*3) {
                             dfs_reset_delayline(&rf->rf_dl);
                             VOS_TRACE(VOS_MODULE_ID_SAP,
